Allow numeric possessives
Example of a very common straight quote in my writing:

    Person 1's grammar is better than person 2's.

smartquotes.js currently handles this incorrectly, replacing the straight quotes with primes:

    Person 1′s grammar is better than person 2′s.

The following is the behavior I expected:

    Person 1’s grammar is better than person 2’s.

The root of the problem is that smartquotes.js was only replacing `'s`
if the previous character matches `[a-z]`.

I think this is too restrictive, and have expanded this to `[a-z0-9]`.
I think there is potential to expand this further.

This change does not break any existing tests.
I have added an additional test for the behavior I wanted.